Separate paths into individual files for existing test cases
